Add page for channel discussion
Add page layout for channel discussion. Create `ThreadHeader`
component that shows Channel name and takes some actions which
display on the right side. Add Thread Message component which
shows the actual body of single message and includes user name,
time since the message was posted, message and actions to take
on the messages.

On Thread page added dummy data so that can be replaced with
actual information

Refer #19